8.4.0.3 (#4)
This branch was more procedural. The 8.4.0.2 branch was created in a seperate process than 8.4.0.2 running on valcor. To ensure all things are the sam I changed the policy. Now a platform must be added to this repo and pulled to valcor. This ensures same same. I created this 8.4.0.3 branch in this manner. The code is the same as 8.4.0.2